Implement `Int64` as a wrapper for `int` when targeting native and Wasm
This PR adds another `Int64` implementation that is just a wrapper
around `int`, to be used in targets where `int` is 64 bits.

Becuase `dart doc` generates using the native class, all documentation
comments are moved to the native class.

The conditional import is implemented based on availability of the
library `dart:html`, which is only available on dart2js.

Improves #172. It could be further improved by having `Int64` as an
extension class when targeting native and Wasm, but that would be a
breaking change.
