+++ pkgs/fixnum/lib/fixnum.dart
@@ -5,8 +5,9 @@
 /// Signed 32- and 64-bit integer support.
 ///
 /// The integer implementations in this library are designed to work
-/// identically whether executed on the Dart VM or compiled to JavaScript.
-library;
+/// identically whether executed on the Dart VM or compiled to JavaScript or
+/// Wasm.
+library fixnum;
 
 export 'src/int32.dart';
 export 'src/int64.dart';
