+++ pkgs/fixnum/lib/src/int32.dart
@@ -321,8 +321,6 @@ class Int32 implements IntX {
     return Int32(value);
   }
 
-  /// Returns [:true:] if this [Int32] has the same numeric value as the
-  /// given object.  The argument may be an [int] or an [IntX].
   @override
   bool operator ==(Object other) {
     if (other is Int32) {
