+++ pkgs/fixnum/test/int_64_vm_test.dart
@@ -3,7 +3,6 @@
 // BSD-style license that can be found in the LICENSE file.
 
 @TestOn('vm')
-library;
 
 import 'package:fixnum/fixnum.dart';
 import 'package:test/test.dart';
@@ -11,42 +10,24 @@ import 'package:test/test.dart';
 void main() {
   group('conversions', () {
     test('toInt', () {
-      expect(
-        Int64.parseInt('-10000000000000000').toInt(),
-        same(-10000000000000000),
-      );
-      expect(
-        Int64.parseInt('-10000000000000001').toInt(),
-        same(-10000000000000001),
-      );
-      expect(
-        Int64.parseInt('-10000000000000002').toInt(),
-        same(-10000000000000002),
-      );
-      expect(
-        Int64.parseInt('-10000000000000003').toInt(),
-        same(-10000000000000003),
-      );
-      expect(
-        Int64.parseInt('-10000000000000004').toInt(),
-        same(-10000000000000004),
-      );
-      expect(
-        Int64.parseInt('-10000000000000005').toInt(),
-        same(-10000000000000005),
-      );
-      expect(
-        Int64.parseInt('-10000000000000006').toInt(),
-        same(-10000000000000006),
-      );
-      expect(
-        Int64.parseInt('-10000000000000007').toInt(),
-        same(-10000000000000007),
-      );
-      expect(
-        Int64.parseInt('-10000000000000008').toInt(),
-        same(-10000000000000008),
-      );
+      expect(Int64.parseInt('-10000000000000000').toInt(),
+          same(-10000000000000000));
+      expect(Int64.parseInt('-10000000000000001').toInt(),
+          same(-10000000000000001));
+      expect(Int64.parseInt('-10000000000000002').toInt(),
+          same(-10000000000000002));
+      expect(Int64.parseInt('-10000000000000003').toInt(),
+          same(-10000000000000003));
+      expect(Int64.parseInt('-10000000000000004').toInt(),
+          same(-10000000000000004));
+      expect(Int64.parseInt('-10000000000000005').toInt(),
+          same(-10000000000000005));
+      expect(Int64.parseInt('-10000000000000006').toInt(),
+          same(-10000000000000006));
+      expect(Int64.parseInt('-10000000000000007').toInt(),
+          same(-10000000000000007));
+      expect(Int64.parseInt('-10000000000000008').toInt(),
+          same(-10000000000000008));
     });
   });
